Created BookingService and BookingStatus

+++ src/main/java/com/pensionat/customer/repository/CustomerRepository.java
@@ -4,5 +4,5 @@ import com.pensionat.customer.model.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<CustomerEntity, Integer> {
+        extends JpaRepository<CustomerEntity, Long> {
 }
